GEOT-2767
git-svn-id: http://svn.osgeo.org/geotools/trunk@34085 e5c1c795-43da-0310-a71f-fac65c449510

+++ modules/extension/xsd/xsd-core/src/main/java/org/geotools/xs/bindings/XSNegativeIntegerBinding.java
@@ -85,7 +85,7 @@ public class XSNegativeIntegerBinding implements SimpleBinding {
         throws Exception {
         Number number = (Number) value;
 
-        if (number.intValue() == 0) {
+        if (number.longValue() == 0) {
             throw new IllegalArgumentException("negativeInteger value '" + number
                 + "' required to be negative");
         }
