+++ modules/extension/xsd/xsd-core/src/main/java/org/geotools/xs/bindings/XSNonNegativeIntegerBinding.java
@@ -119,7 +119,7 @@ public class XSNonNegativeIntegerBinding implements SimpleBinding {
     public String encode(Object object, String value) throws Exception {
         Number number = (Number) object;
 
-        if (number.intValue() < 0) {
+        if (number.longValue() < 0) {
             throw new IllegalArgumentException("Value '" + number
                 + "' must be non-negative (0 or above).");
         }
